+++ package.json
@@ -22,6 +22,7 @@
 	"homepage": "https://github.com/Gen-x-academy/chainVerse-backend#readme",
 	"dependencies": {
 		"bcrypt": "^5.1.1",
+		"bcryptjs": "^3.0.2",
 		"body-parser": "^1.20.3",
 		"cookie-parser": "^1.4.7",
 		"cors": "^2.8.5",
